create a separate public layer
